Refactor Logic.js and ExpertChatbo
